+++ .gitignore
@@ -12,7 +12,7 @@ coverage
 .next/
 out/
 build
-
+tsbuildinfo
 # misc
 .DS_Store
 *.pem
