+++ answeroverflow.code-workspace
@@ -88,7 +88,7 @@
       "**/coverage/": true,
       "**/.next": true,
       "**/.turbo": true,
-      "**/dist": true,
+      // "**/dist": true,
     },
     // "typescript.tsserver.experimental.enableProjectDiagnostics": true,
     "vitest.enable": true,
@@ -177,6 +177,7 @@
       "tailwindcss",
       "threadable",
       "trpc",
+      "tsup",
       "Turborepo",
       "twoslash",
       "unauthed",
